+++ src/main/cpp/nativeimage/NativeImage.cpp
@@ -4,27 +4,29 @@
 // Node APIs are not fully supported. To solve the compilation error of the interface cannot be found,
 // please include "napi/native_api.h".
 
+#include <EGL/egl.h>
 #include <GLES2/gl2.h>
 #include <GLES2/gl2ext.h>
 #include <cstdio>
 #include <native_image/native_image.h>
-#include <native_window/external_window.h>
 #include "NativeImage.h"
 #include "atomic/Atomic.h"
 #include "common/utils.h"
+#include <hilog/log.h>
 
 
 namespace NativeImage {
 
-std::unordered_map<OH_NativeImage *, std::atomic_int *> atomic_ints;
+std::unordered_map<OH_NativeImage *, napi_threadsafe_function> sf_cache;
 
 void NapiDispose(napi_env env, void *finalize_data, void *finalize_hint) {
-    OH_NativeImage *image = (OH_NativeImage *)finalize_data;
+    OH_NativeImage *image = static_cast<OH_NativeImage *>(finalize_data);
     if (image == nullptr)
         return;
+
     OH_NativeImage_UnsetOnFrameAvailableListener(image);
     OH_NativeImage_Destroy(&image);
-    atomic_ints.erase((OH_NativeImage *)image);
+    DestroyThreadSafeFunction(env, image);
     image = nullptr;
 }
 
@@ -77,36 +79,25 @@ napi_value NapiBindNativeImage(napi_env env, napi_callback_info info) {
     if (image == nullptr)
         napi_throw_error(env, "NativeImage", "Create NativeImage Failed");
 
-    OH_OnFrameAvailableListener listener;
-    listener.context = flag;
-    listener.onFrameAvailable = [](void *context) {
-        std::atomic_int *flag = (std::atomic_int *)context;
-        (*flag)++;
-    };
+//     OH_OnFrameAvailableListener listener;
+//     listener.context = flag;
+//     listener.onFrameAvailable = [](void *context) {
+//         std::atomic_int *flag = (std::atomic_int *)context;
+//         (*flag)++;
+//     };
 
-    int error = OH_NativeImage_SetOnFrameAvailableListener(image, listener);
-    if (error != 0)
-        napi_throw_error(env, "NativeImage", "Set FrameAvailableListener Failed");
+//     int error = OH_NativeImage_SetOnFrameAvailableListener(image, listener);
+//     if (error != 0)
+//         napi_throw_error(env, "NativeImage", "Set FrameAvailableListener Failed");
 
-    napi_wrap_sendable(
-        env, _this, static_cast<void *>(image),
-        [](napi_env env, void *data, void *hint) {
-            OH_NativeImage *image = (OH_NativeImage *)data;
-            if (image == nullptr)
-                return;
-            OH_NativeImage_UnsetOnFrameAvailableListener(image);
-            OH_NativeImage_Destroy(&image);
-            atomic_ints.erase((OH_NativeImage *)image);
-            image = nullptr;
-        },
-        nullptr);
+    napi_wrap_sendable(env, _this, static_cast<void *>(image), NapiDispose, nullptr);
 
-    atomic_ints[image] = flag;
     return _this;
 }
 
 napi_value NapiUpdateSurfaceImage(napi_env env, napi_callback_info info) {
 
+    auto id = std::this_thread::get_id();
     napi_value _this = nullptr;
     napi_get_cb_info(env, info, nullptr, nullptr, &_this, nullptr);
 
@@ -116,11 +107,7 @@ napi_value NapiUpdateSurfaceImage(napi_env env, napi_callback_info info) {
         napi_throw_error(env, "NativeImage::UpdateSurfaceImage", "invalid operation");
         return nullptr;
     }
-
-    auto it = atomic_ints.find(image);
-    if (it == atomic_ints.end())
-        napi_throw_error(env, "NativeImage", "Unknown problem");
-    (*(it->second))--;
+    
     int error = OH_NativeImage_UpdateSurfaceImage((OH_NativeImage *)image);
     return createNapiInt32(env, error);
 }
@@ -131,7 +118,7 @@ napi_value NapiDestroyNativeImage(napi_env env, napi_callback_info info) {
 
     void *image = nullptr;
     napi_remove_wrap_sendable(env, _this, &image);
-    atomic_ints.erase((OH_NativeImage *)image);
+
     return nullptr;
 }
 
@@ -157,22 +144,82 @@ napi_value NapiGetIsAvailable(napi_env env, napi_callback_info info) {
     if (image == nullptr)
         napi_throw_error(env, "NativeImage", "Get Available Failed, NativeImage IS Null");
 
-    auto it = atomic_ints.find(image);
-    if (it != atomic_ints.end()) {
-        int val = it->second->load(std::memory_order_relaxed);
-        napi_value result = nullptr;
-        napi_create_int32(env, val, &result);
-        return result;
-    } else {
-        napi_throw_error(env, "NativeImage", "Get Available Failed, Atomic_int IS Null");
-        return nullptr;
-    }
+
 }
 
 OH_NativeImage *GetNativeImage(napi_env env, napi_value value) {
     void *result = nullptr;
     napi_unwrap_sendable(env, value, &result);
     return static_cast<OH_NativeImage *>(result);
+}
+
+napi_value NapiSetOnFrameAvailableListener(napi_env env, napi_callback_info info) {
+    size_t argc = 1;
+    napi_value argv[1];
+    napi_value _this = nullptr;
+    napi_get_cb_info(env, info, &argc, argv, &_this, nullptr);
+
+    if (argc != 1)
+        napi_throw_error(env, "NativeImage", "SetOnFrameAvailableListener must got 1 argument");
+
+    OH_NativeImage *image = GetNativeImage(env, _this);
+    if (image == nullptr) {
+        napi_throw_error(env, "NativeImage", "SetOnFrameAvailableListener Failed, NativeImage IS Null");
+        return nullptr;
+    }
+
+    napi_valuetype type;
+    napi_typeof(env, argv[0], &type);
+
+    if (type != napi_function) {
+        napi_throw_error(env, "NativeImage", "SetOnFrameAvailableListener: arg 0 not a function");
+        return nullptr;
+    }
+
+//     FrameListener *onFrame = new FrameListener();
+
+    napi_threadsafe_function sf = nullptr;
+    napi_value resourceName = nullptr;
+    napi_create_string_utf8(env, "onFrameCallback", NAPI_AUTO_LENGTH, &resourceName);
+    napi_create_threadsafe_function(env, argv[0], nullptr, resourceName, 0, 1, nullptr, nullptr, nullptr, RaiseFrameAvailable, &sf);
+
+
+    OH_OnFrameAvailableListener listener;
+    listener.context = sf;
+    listener.onFrameAvailable = [](void *context) {
+        napi_threadsafe_function func = static_cast<napi_threadsafe_function>(context);
+        napi_acquire_threadsafe_function(func);
+        napi_call_threadsafe_function(func, nullptr, napi_tsfn_nonblocking);
+    };
+    auto result = OH_NativeImage_SetOnFrameAvailableListener(image, listener);
+    sf_cache[image] = sf;
+    return createNapiInt32(env, result);
+}
+
+napi_value NapiUnSetOnFrameAvailableListener(napi_env env, napi_callback_info info) {
+    napi_value _this = nullptr;
+    napi_get_cb_info(env, info, nullptr, nullptr, &_this, nullptr);
+    OH_NativeImage *image = GetNativeImage(env, _this);
+    if (image == nullptr) {
+        napi_throw_error(env, "NativeImage", "UnSetOnFrameAvailableListener: NativeImage IS Null");
+        return nullptr;
+    }
+
+    auto result = OH_NativeImage_UnsetOnFrameAvailableListener(image);
+    DestroyThreadSafeFunction(env, image);
+    return createNapiInt32(env, result);
+}
+
+void RaiseFrameAvailable(napi_env env, napi_value js_callback, void *context, void *data) {
+    napi_call_function(env, nullptr, js_callback, 0, nullptr, nullptr);
+}
+
+void DestroyThreadSafeFunction(napi_env env, OH_NativeImage *image) {
+    auto it = sf_cache.find(image);
+    if (it != sf_cache.end()) {
+        napi_release_threadsafe_function(it->second, napi_tsfn_release);
+        sf_cache.erase(it);
+    }
 }
 
 } // namespace NativeImage