+++ src/main/cpp/nativeimage/NativeImage.h
@@ -12,6 +12,11 @@
 #include <unordered_map>
 
 namespace NativeImage {
+
+struct FrameListener {
+    napi_threadsafe_function sf;
+};
+
 // napi_value NapiConstructor(napi_env env, napi_callback_info info);
 void NapiDispose(napi_env env, void *finalize_data, void *finalize_hint);
 napi_value NapiAttachContext(napi_env env, napi_callback_info info);
@@ -21,7 +26,11 @@ napi_value NapiBindNativeImage(napi_env env, napi_callback_info info);
 napi_value NapiDestroyNativeImage(napi_env env, napi_callback_info info);
 napi_value NapiGetSurfaceId(napi_env env, napi_callback_info info);
 napi_value NapiGetIsAvailable(napi_env env, napi_callback_info info);
-OH_NativeImage* GetNativeImage(napi_env env, napi_value value);
+OH_NativeImage *GetNativeImage(napi_env env, napi_value value);
+napi_value NapiSetOnFrameAvailableListener(napi_env env, napi_callback_info info);
+napi_value NapiUnSetOnFrameAvailableListener(napi_env env, napi_callback_info info);
+void RaiseFrameAvailable(napi_env env, napi_value js_callback, void *context, void *data);
+void DestroyThreadSafeFunction(napi_env env, OH_NativeImage *image);
 } // namespace NativeImage
 
 #endif // GLTEST_NATIVEIMAGE_H